Port generate equals and hash code tests to the new framework
